+++ src/Dobee/Http/Cookie/CookieInterface.php
@@ -15,7 +15,8 @@ namespace Dobee\Http\Cookie;
 
 /**
  * Interface CookieInterface
- * @package Dobee\Http
+ *
+ * @package Dobee\Http\Cookie
  */
 interface CookieInterface
 {
@@ -30,27 +31,74 @@ interface CookieInterface
      */
     public function getName();
 
+    /**
+     * @param $value
+     * @return $this
+     */
     public function setValue($value);
 
+    /**
+     * @return string
+     */
     public function getValue();
 
+    /**
+     * @param $time
+     * @return $this
+     */
     public function setExpire($time);
 
+    /**
+     * @return int
+     */
     public function getExpire();
 
+    /**
+     * @param string $path
+     * @return $this
+     */
     public function setPath($path = '/');
 
+    /**
+     * @return string
+     */
     public function getPath();
 
+    /**
+     * @param string $domain
+     * @return $this
+     */
     public function setDomain($domain);
 
+    /**
+     * @return string
+     */
     public function getDomain();
 
+    /**
+     * @param bool $secure
+     * @return $this
+     */
     public function setSecure($secure = false);
 
+    /**
+     * @return bool
+     */
     public function getSecure();
 
+    /**
+     * @param bool $httpOnly
+     * @return $this
+     */
     public function setHttpOnly($httpOnly = false);
 
+    /**
+     * @return bool
+     */
     public function getHttpOnly();
+
+    /**
+     * @return mixed
+     */
+    public function clear();
 }